DARK MODE ✨🌑 Update minimal-mistakes skin imports and dark mode stylesheet
This commit updates the import statements for the minimal-mistakes skin in the main.scss file. It replaces the variable `site.minimal_mistakes_skin` with `site.mm_skin` for the default skin. Additionally, a new import statement for the dark mode stylesheet is added in the main-dark.scss file. This allows the website to switch between light and dark themes based on the user's preferred color scheme.